refactor: move ResponseWithPagination to common/dto
