+++ src/main/java/com/ajou/travely/controller/user/UserController.java
@@ -2,7 +2,7 @@ package com.ajou.travely.controller.user;
 
 import com.ajou.travely.config.auth.LoginUser;
 import com.ajou.travely.config.auth.SessionUser;
-import com.ajou.travely.controller.travel.dto.ResponseWithPagination;
+import com.ajou.travely.controller.common.ResponseWithPagination;
 import com.ajou.travely.controller.travel.dto.SimpleTravelResponseDto;
 import com.ajou.travely.controller.user.dto.UserCreateRequestDto;
 import com.ajou.travely.controller.user.dto.UserResponseDto;
